Reorganize input components
