+++ client/src/components/input/numeric/GBP/GBP.tsx
