+++ client/src/components/input/numeric/GBP/gbp.test.jsx
