+++ client/src/components/input/sortSelector/sortSelector.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useSortContext } from '../../contexts/sort';
 import Button from '../../buttons/secondaryButton';
-import Select from './select';
+import Select from '../select/select';
 
 function SortSelector (): JSX.Element {
